+++ packages/@eventual/core-runtime/src/providers/entity-provider.ts
@@ -1,4 +1,4 @@
-import type { AnyEntity } from "@eventual/core";
+import type { Entity } from "@eventual/core";
 import { entities } from "@eventual/core/internal";
 import type { WorkflowExecutor } from "../workflow-executor.js";
 
@@ -8,14 +8,14 @@ export interface EntityProvider {
    *
    * Use {@link WorkflowExecutor}.isStarted to determine if it is already started.
    */
-  getEntity(entityName: string): AnyEntity | undefined;
+  getEntity(entityName: string): Entity | undefined;
 }
 
 /**
  * An executor provider that works with an out of memory store.
  */
 export class GlobalEntityProvider implements EntityProvider {
-  public getEntity(entityName: string): AnyEntity | undefined {
+  public getEntity(entityName: string): Entity | undefined {
     return entities().get(entityName);
   }
 }
