+++ packages/@eventual/core/src/index.ts
@@ -1,7 +1,7 @@
 export * from "./await-time.js";
 export * from "./bucket.js";
 export * from "./condition.js";
-export * from "./entity.js";
+export * from "./entity/index.js";
 export * from "./error.js";
 export * from "./event.js";
 export * from "./execution.js";
